test (search v2): make object with R2R utility functions
