+++ webapi/src/test/scala/org/knora/webapi/responders/v2/SearchResponderV2Spec.scala
@@ -26,7 +26,7 @@ import org.knora.webapi.messages.store.triplestoremessages.{RdfDataObject, Reset
 import org.knora.webapi.messages.v1.responder.ontologymessages.{LoadOntologiesRequest, LoadOntologiesResponse}
 import org.knora.webapi.messages.v2.responder.ReadResourcesSequenceV2
 import org.knora.webapi.messages.v2.responder.searchmessages._
-import org.knora.webapi.responders.v2.ResponseCheckerV2.compareReadResourcesSequenceV2Response
+import org.knora.webapi.responders.v2.ResponseCheckerResponderV2.compareReadResourcesSequenceV2Response
 import org.knora.webapi.responders.{RESPONDER_MANAGER_ACTOR_NAME, ResponderManager}
 import org.knora.webapi.store.{STORE_MANAGER_ACTOR_NAME, StoreManager}
 import org.knora.webapi.util.MessageUtil
